Move some of the layout helpers into concerns
Move all the other page layout helpers

Move the building page layout methods

Move classes out of layout helper

Fix issues with rubocop

Rename all page detail helpers

Update the buildings concern
