+++ app/controllers/concerns/facilities_management/page_detail/supplier/dashboard.rb
@@ -1,0 +1,13 @@
+module FacilitiesManagement::PageDetail::Supplier::Dashboard
+  include FacilitiesManagement::PageDetail
+
+  def page_definitions
+    @page_definitions ||= {
+      default: {},
+      index: {
+        page_title: 'Direct award dashboard',
+        caption1: current_user.email
+      }
+    }.freeze
+  end
+end
