+++ app/controllers/facilities_management/procurements/edit_buildings_controller.rb
@@ -1,41 +1,16 @@
 module FacilitiesManagement
   module Procurements
     class EditBuildingsController < FacilitiesManagement::FrameworkController
-      include Procurements::EditBuildingsControllerDefinitions
       include Procurements::EditBuildingsControllerNavigation
+      include FacilitiesManagement::PageDetail::EditBuildings
       include FacilitiesManagement::FindAddressConcern
-      include SharedBuildingsControllerMethods
-
-      before_action :set_procurement
-      before_action :initialise_page_data
-      before_action :create_new_building, only: :new
-      before_action :build_page_data, only: %i[show edit update add_address]
-      before_action :redirect_if_unrecognised_step, only: :edit
-      before_action :authorize_user
-      before_action :build_page_description, only: %i[show new edit add_address]
-      before_action :initialize_building_details, only: :show
-
-      def new; end
-
-      def add_address; end
-
-      def show; end
-
-      def edit; end
-
-      def create
-        create_building_action
-      end
-
-      def update
-        update_building_action
-      end
+      include BuildingsConcern
 
       private
 
       def add_address_form_details
         @add_address_form_details ||= if id_present?
-                                        { url: facilities_management_procurement_edit_building_path(@page_data[:model_object].id, procurement_id: @procurement.id), method: :patch }
+                                        { url: facilities_management_procurement_edit_building_path(@page_data[:model_object].id, procurement_id: params[:procurement_id]), method: :patch }
                                       else
                                         { url: facilities_management_procurement_edit_buildings_path, method: :post }
                                       end
@@ -43,12 +18,8 @@ module FacilitiesManagement
 
       helper_method :step_title, :step_footer, :add_address_form_details, :valid_regions, :region_needs_resolution?, :multiple_regions?
 
-      def set_procurement
-        @procurement = current_user.procurements.find(params[:procurement_id])
-      end
-
       def redirect_if_unrecognised_step
-        redirect_to facilities_management_procurement_edit_building_path(@procurement, @page_data[:model_object]) unless RECOGNISED_STEPS.include? params[:step]
+        redirect_to facilities_management_procurement_edit_building_path(params[:procurement_id], @page_data[:model_object]) unless RECOGNISED_STEPS.include? params[:step]
       end
 
       protected
