+++ app/helpers/layout_helper.rb
@@ -2,75 +2,10 @@
 module LayoutHelper
   include ErrorsHelper
 
-  # Module to render page elements
-  # field-sets and input groups
-  # errors
-
-  # Value Objects (Classes) to structure data for parameters to helper methods
-  class NavigationDetail
-    attr_accessor(:primary_text, :primary_name, :primary_url, :return_url, :return_text, :secondary_url, :secondary_text, :secondary_name)
-
-    # rubocop:disable Metrics/ParameterLists
-    def initialize(primary_text, return_url, return_text, secondary_url, secondary_text, primary_name = nil, secondary_name = nil, primary_url = nil)
-      @primary_text   = primary_text
-      @primary_name   = primary_name
-      @primary_url = primary_url
-      @return_url     = return_url
-      @return_text    = return_text
-      @secondary_url  = secondary_url
-      @secondary_text = secondary_text
-      @secondary_name = secondary_name
-    end
-    # rubocop:enable Metrics/ParameterLists
-  end
-
-  class BackButtonDetail
-    attr_accessor(:url, :text, :label)
-
-    def initialize(back_url, back_label, back_text)
-      @url   = back_url
-      @label = back_label
-      @text  = back_text
-    end
-  end
-
-  class HeadingDetail
-    attr_accessor(:text, :caption, :caption2, :subtitle, :caption3)
-
-    def initialize(header_text, caption1, caption2, sub_text, caption3)
-      @text     = header_text
-      @caption  = caption1
-      @caption2 = caption2
-      @subtitle = sub_text
-      @caption3 = caption3
-    end
-
-    def caption?
-      @caption.present? || caption2.present? || caption3.present?
-    end
-  end
-
-  class PageDescription
-    attr_accessor :heading_details, :back_button, :navigation_details, :no_back_button, :no_error_block, :no_headings
-
-    def initialize(heading_details, back_button = nil, continuation = nil)
-      raise ArgumentError, 'Use a HeadingDetails object' unless heading_details.is_a? HeadingDetail
-
-      raise ArgumentError, 'Use a BackButtonDetail object' unless back_button.nil? || back_button.is_a?(BackButtonDetail)
-
-      raise ArgumentError, 'Use a NavigationDetail object' unless continuation.nil? || continuation.is_a?(NavigationDetail)
-
-      @no_back_button     = @no_error_block = @no_headings = false
-      @heading_details    = heading_details
-      @back_button        = back_button
-      @navigation_details = continuation
-    end
-  end
-
   # Renders the top of the page including back-button, and the 3 elements of the main header
   # rubocop:disable Rails/OutputSafety, Metrics/ParameterLists
   def govuk_page_content(page_details, model_object = nil, no_headings = false, no_back_button = false, no_error_block = false)
-    raise ArgumentError, 'Use PageDescription object' unless page_details.is_a? PageDescription
+    raise ArgumentError, 'Use PageDescription object' unless page_details.is_a? FacilitiesManagement::PageDetail::PageDescription
 
     @no_back_button = no_back_button
     @no_error_block = no_error_block
