+++ app/views/crown_marketplace/allow_list/new.html.erb
@@ -5,7 +5,7 @@
 <%= form_with url: crown_marketplace_allow_list_index_path, model: @allowed_email_domain, method: :post, local: true, html: { specialvalidation: true } do |f| %>
   <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
-      <%= govuk_page_header(LayoutHelper::HeadingDetail.new(t('.heading'), nil, nil, nil, nil)) %>
+      <%= govuk_page_header(FacilitiesManagement::PageDetail::Heading.new(t('.heading'), nil, nil, nil, nil)) %>
 
       <%= form_group_with_error f.object, :email_domain do |displayed_error| %>
         <%= label_details(f, :email_domain, t('.email_domain_label'), t('.email_domain_hint')) %>
