+++ app/views/facilities_management/admin/supplier_details/edit.html.erb
@@ -7,7 +7,7 @@
 
 <%= form_with model: @supplier, url: facilities_management_admin_supplier_detail_path(page: @page), local: 'false', html: { specialvalidation: true } do |f| %>
   <%= render partial: 'shared/error_summary', locals: { errors: f.object.errors } %>
-  <%= govuk_page_header(LayoutHelper::HeadingDetail.new(t(".page_title.#{@page}"), current_supplier_name, nil, nil, nil)) %>
+  <%= govuk_page_header(FacilitiesManagement::PageDetail::Heading.new(t(".page_title.#{@page}"), current_supplier_name, nil, nil, nil)) %>
   <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
       <%= render partial: @page.to_s, locals: { f: f } %>
