+++ app/views/facilities_management/admin/supplier_details/show.html.erb
@@ -3,7 +3,7 @@
   { link_text: t('facilities_management.admin.supplier_details.index.heading'), link_url: facilities_management_admin_supplier_details_path },
   { link_text: @supplier.supplier_name }
 ) %>
-<%= govuk_page_header(LayoutHelper::HeadingDetail.new(t('.supplier_details'), @supplier.supplier_name, nil, nil, nil)) %>
+<%= govuk_page_header(FacilitiesManagement::PageDetail::Heading.new(t('.supplier_details'), @supplier.supplier_name, nil, nil, nil)) %>
 
 <p class="govuk-body govuk-!-font-weight-regular govuk-hint">
   <%= t('.edit_supplier_details') %>
