+++ app/views/facilities_management/procurement_buildings/_buildings_and_services.html.erb
@@ -4,7 +4,7 @@
   <div class="govuk-grid-column-two-thirds">
     <%= link_to t('.return_link'), facilities_management_procurement_summary_path(@procurement, summary: 'buildings_and_services'),  'aria-label': t('.return_link'), class: 'govuk-back-link govuk-!-margin-top-0 govuk-!-margin-bottom-6' %>
     <%= render partial: 'shared/error_summary', locals: { errors: @procurement_building.errors } %>
-    <%= govuk_page_header(LayoutHelper::HeadingDetail.new(@procurement_building.name, @procurement.contract_name, nil, nil, nil)) %>
+    <%= govuk_page_header(FacilitiesManagement::PageDetail::Heading.new(@procurement_building.name, @procurement.contract_name, nil, nil, nil)) %>
   </div>
 </div>
 <%= form_group_with_error(@procurement_building, :service_codes) do |displayed_error| %>
