+++ app/views/facilities_management/procurements/edit/_buildings.html.erb
@@ -1,5 +1,5 @@
 <div class="govuk-!-width-three-quarters">
-  <%= govuk_page_header(LayoutHelper::HeadingDetail.new(t('.heading'), @procurement.contract_name, nil, nil, nil)) %>
+  <%= govuk_page_header(FacilitiesManagement::PageDetail::Heading.new(t('.heading'), @procurement.contract_name, nil, nil, nil)) %>
 </div>
 <%= content_for :page_title, t(".heading") %>
 <div class="govuk-grid-row">
