+++ app/views/facilities_management/procurements/edit/_contract_period.html.erb
@@ -1,5 +1,5 @@
 <div class="govuk-form-group">
-  <%= govuk_page_header(LayoutHelper::HeadingDetail.new(t('.title'), @procurement.contract_name, nil, nil, nil)) %>
+  <%= govuk_page_header(FacilitiesManagement::PageDetail::Heading.new(t('.title'), @procurement.contract_name, nil, nil, nil)) %>
   <% content_for :page_title, t('.title') %>
   <div class="contract-dates">
     <%= render partial: 'facilities_management/procurements/edit/contract_period/initial_call_off_period', locals: { f: f } %>
