+++ app/views/facilities_management/procurements/edit_buildings/show.html.erb
@@ -1,1 +1,1 @@
-<%= render partial: 'facilities_management/buildings/action_partials/show', locals: { return_link_text: t('.return_to_buildings'), return_link: edit_facilities_management_procurement_path(@procurement.id, step: 'buildings') }%>
+<%= render partial: 'facilities_management/buildings/action_partials/show', locals: { return_link_text: t('.return_to_buildings'), return_link: edit_facilities_management_procurement_path(params[:procurement_id], step: 'buildings') }%>
