+++ app/views/facilities_management/procurements/spreadsheet_imports/new.html.erb
@@ -3,7 +3,7 @@
     <%= render partial: 'shared/error_summary', locals: { errors: f.object.errors } %>
   <% end %>
 
-  <%= govuk_page_header(LayoutHelper::HeadingDetail.new(t('.heading'), @spreadsheet_import.procurement.contract_name, nil, nil, nil)) %>
+  <%= govuk_page_header(FacilitiesManagement::PageDetail::Heading.new(t('.heading'), @spreadsheet_import.procurement.contract_name, nil, nil, nil)) %>
   <% content_for :page_title, t('.heading') %>
 
   <div class="govuk-body govuk-!-width-two-thirds">
