+++ app/views/facilities_management/procurements/spreadsheet_imports/show.html.erb
@@ -1,6 +1,6 @@
 <%= link_to t('.return_to_upload'), new_facilities_management_procurement_spreadsheet_import_path(@spreadsheet_import.procurement), class: 'govuk-back-link govuk-!-margin-bottom-6' %>
 
-<%= govuk_page_header(LayoutHelper::HeadingDetail.new(t('.heading'), @procurement.contract_name, nil, nil, nil)) %>
+<%= govuk_page_header(FacilitiesManagement::PageDetail::Heading.new(t('.heading'), @procurement.contract_name, nil, nil, nil)) %>
 <% content_for :page_title, "#{t('.heading')}, #{@spreadsheet_import.state_to_string[1].downcase}"  %>
 
 <div class="ccs-font-weight-semi-bold govuk-body">
