+++ app/views/facilities_management/procurements/summary.html.erb
@@ -1,4 +1,4 @@
-<%= govuk_page_header(LayoutHelper::HeadingDetail.new(t(".heading.#{@summary_page}"), @procurement.contract_name, nil, nil, nil)) %>
+<%= govuk_page_header(FacilitiesManagement::PageDetail::Heading.new(t(".heading.#{@summary_page}"), @procurement.contract_name, nil, nil, nil)) %>
 <%= content_for :page_title, t(".heading.#{@summary_page}") %>
 <%= render "facilities_management/procurements/summary/#{@summary_page}" %>
 <%= link_to t('.return_to_requirements'), facilities_management_procurement_path(@procurement), class: 'govuk-link govuk-link--no-visited-state govuk-!-font-size-19'%>
